# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Forge API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/operations": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "List Operations",
        "description": "List operations with current selection, risk labels, and user overrides.",
        "operationId": "list_operations_api_projects__project_id__operations_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method"
            }
          },
          {
            "name": "risk",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Operations"
        ],
        "summary": "Bulk Update Operations",
        "description": "Bulk update operation enabled state, tool names, descriptions, or groups.",
        "operationId": "bulk_update_operations_api_projects__project_id__operations_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationsBulkUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Bulk Update Operations Api Projects  Project Id  Operations Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/operations/preset": {
      "post": {
        "tags": [
          "Operations"
        ],
        "summary": "Apply Operations Preset",
        "description": "Apply selection preset across all operations (read-only, all, none, by-tag).",
        "operationId": "apply_operations_preset_api_projects__project_id__operations_preset_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationsPresetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Apply Operations Preset Api Projects  Project Id  Operations Preset Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationItemResponse": {
        "properties": {
          "operation_key": {
            "type": "string",
            "title": "Operation Key"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "deprecated": {
            "type": "boolean",
            "title": "Deprecated",
            "default": false
          },
          "risk": {
            "type": "string",
            "enum": [
              "read",
              "write",
              "destructive"
            ],
            "title": "Risk"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "skipped": {
            "type": "boolean",
            "title": "Skipped",
            "default": false
          },
          "skipped_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skipped Reason"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "annotations": {
            "additionalProperties": true,
            "type": "object",
            "title": "Annotations"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          }
        },
        "type": "object",
        "required": [
          "operation_key",
          "method",
          "path",
          "risk",
          "enabled",
          "tool_name"
        ],
        "title": "OperationItemResponse"
      },
      "OperationUpdateItem": {
        "properties": {
          "operation_key": {
            "type": "string",
            "title": "Operation Key"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "tool_name_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name Override"
          },
          "description_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Override"
          },
          "group_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Override"
          }
        },
        "type": "object",
        "required": [
          "operation_key"
        ],
        "title": "OperationUpdateItem"
      },
      "OperationsBulkUpdateRequest": {
        "properties": {
          "operations": {
            "items": {
              "$ref": "#/components/schemas/OperationUpdateItem"
            },
            "type": "array",
            "title": "Operations"
          }
        },
        "type": "object",
        "required": [
          "operations"
        ],
        "title": "OperationsBulkUpdateRequest"
      },
      "OperationsListResponse": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "enabled_count": {
            "type": "integer",
            "title": "Enabled Count"
          },
          "read_count": {
            "type": "integer",
            "title": "Read Count"
          },
          "write_count": {
            "type": "integer",
            "title": "Write Count"
          },
          "destructive_count": {
            "type": "integer",
            "title": "Destructive Count"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/OperationItemResponse"
            },
            "type": "array",
            "title": "Operations"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "enabled_count",
          "read_count",
          "write_count",
          "destructive_count",
          "operations"
        ],
        "title": "OperationsListResponse"
      },
      "OperationsPresetRequest": {
        "properties": {
          "preset": {
            "type": "string",
            "enum": [
              "read-only",
              "all",
              "none",
              "by-tag"
            ],
            "title": "Preset"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "preset"
        ],
        "title": "OperationsPresetRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}